Refactor NodeAggregateCommandHandler to feature traits
