+++ Neos.EventSourcedContentRepository/Classes/Domain/Context/NodeAggregate/Feature/NodeModification.php
@@ -1,0 +1,65 @@
+<?php
+declare(strict_types=1);
+namespace Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate\Feature;
+
+/*
+ * This file is part of the Neos.ContentRepository package.
+ *
+ * (c) Contributors of the Neos Project - www.neos.io
+ *
+ * This package is Open Source Software. For the full copyright and license
+ * information, please view the LICENSE file which was distributed with this
+ * source code.
+ */
+
+use Neos\EventSourcedContentRepository\Domain\Context\ContentStream\ContentStreamEventStreamName;
+use Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate\Command\SetNodeProperties;
+use Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate\Event\NodePropertiesWereSet;
+use Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate\NodeAggregateEventPublisher;
+use Neos\EventSourcedContentRepository\Domain\ValueObject\CommandResult;
+use Neos\EventSourcedContentRepository\Service\Infrastructure\ReadSideMemoryCacheManager;
+use Neos\EventSourcing\Event\Decorator\EventWithIdentifier;
+use Neos\EventSourcing\Event\DomainEvents;
+
+trait NodeModification
+{
+    abstract protected function getReadSideMemoryCacheManager(): ReadSideMemoryCacheManager;
+
+    abstract protected function getNodeAggregateEventPublisher(): NodeAggregateEventPublisher;
+
+    /**
+     * @param SetNodeProperties $command
+     * @return CommandResult
+     */
+    public function handleSetNodeProperties(SetNodeProperties $command): CommandResult
+    {
+        $this->getReadSideMemoryCacheManager()->disableCache();
+
+        $events = null;
+        $this->getNodeAggregateEventPublisher()->withCommand($command, function () use ($command, &$events) {
+            $contentStreamIdentifier = $command->getContentStreamIdentifier();
+
+            // Check if node exists
+            // @todo: this must also work when creating a copy on write
+            #$this->assertNodeWithOriginDimensionSpacePointExists($contentStreamIdentifier, $command->getNodeAggregateIdentifier(), $command->getOriginDimensionSpacePoint());
+
+            $events = DomainEvents::withSingleEvent(
+                EventWithIdentifier::create(
+                    new NodePropertiesWereSet(
+                        $contentStreamIdentifier,
+                        $command->getNodeAggregateIdentifier(),
+                        $command->getOriginDimensionSpacePoint(),
+                        $command->getPropertyValues()
+                    )
+                )
+            );
+
+            $this->getNodeAggregateEventPublisher()->publishMany(
+                ContentStreamEventStreamName::fromContentStreamIdentifier($contentStreamIdentifier)->getEventStreamName(),
+                $events
+            );
+        });
+
+        return CommandResult::fromPublishedEvents($events);
+    }
+}
