+++ Neos.EventSourcedContentRepository/Classes/Domain/Context/NodeAggregate/Feature/NodeRenaming.php
@@ -1,0 +1,61 @@
+<?php
+declare(strict_types=1);
+namespace Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate\Feature;
+
+/*
+ * This file is part of the Neos.ContentRepository package.
+ *
+ * (c) Contributors of the Neos Project - www.neos.io
+ *
+ * This package is Open Source Software. For the full copyright and license
+ * information, please view the LICENSE file which was distributed with this
+ * source code.
+ */
+
+use Neos\EventSourcedContentRepository\Domain\Context\ContentStream\ContentStreamEventStreamName;
+use Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate\Command\ChangeNodeAggregateName;
+use Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate\Event\NodeAggregateNameWasChanged;
+use Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate\NodeAggregateEventPublisher;
+use Neos\EventSourcedContentRepository\Domain\ValueObject\CommandResult;
+use Neos\EventSourcedContentRepository\Service\Infrastructure\ReadSideMemoryCacheManager;
+use Neos\EventSourcing\Event\Decorator\EventWithIdentifier;
+use Neos\EventSourcing\Event\DomainEvents;
+
+trait NodeRenaming
+{
+    abstract protected function getReadSideMemoryCacheManager(): ReadSideMemoryCacheManager;
+
+    abstract protected function getNodeAggregateEventPublisher(): NodeAggregateEventPublisher;
+
+    /**
+     * @param ChangeNodeAggregateName $command
+     * @return CommandResult
+     */
+    public function handleChangeNodeAggregateName(ChangeNodeAggregateName $command): CommandResult
+    {
+        // TODO: check if CS exists
+        // TODO: check if aggregate exists and delegate to it
+        // TODO: check if aggregate is root
+        $events = DomainEvents::fromArray([]);
+        $this->getNodeAggregateEventPublisher()->withCommand($command, function () use ($command, &$events) {
+            $events = DomainEvents::withSingleEvent(
+                EventWithIdentifier::create(
+                    new NodeAggregateNameWasChanged(
+                        $command->getContentStreamIdentifier(),
+                        $command->getNodeAggregateIdentifier(),
+                        $command->getNewNodeName()
+                    )
+                )
+            );
+
+            $this->getNodeAggregateEventPublisher()->publishMany(
+                ContentStreamEventStreamName::fromContentStreamIdentifier(
+                    $command->getContentStreamIdentifier()
+                )->getEventStreamName(),
+                $events
+            );
+        });
+
+        return CommandResult::fromPublishedEvents($events);
+    }
+}
